// convex/_generated/api.d.ts
// This file is auto-generated by Convex.
// Run `npx convex dev` to generate the actual types.

declare const api: {
  products: {
    getAll: any;
    getBySlug: any;
    getBestSellers: any;
    getFeatured: any;
    search: any;
    getByCategory: any;
    create: any;
    update: any;
    remove: any;
  };
  orders: {
    create: any;
    get: any;
    getByOrderNumber: any;
    getByEmail: any;
    getByUserId: any;
    getAll: any;
    list: any;
    updatePaymentStatus: any;
    updateStatus: any;
    updateFulfillmentStatus: any;
    addInternalNote: any;
    getRecentOrders: any;
    getOrderStats: any;
  };
  cart: {
    get: any;
    getBySession: any;
    addItem: any;
    updateItemQuantity: any;
    removeItem: any;
    applyDiscount: any;
    removeDiscount: any;
    clear: any;
    merge: any;
  };
  discountCodes: {
    validate: any;
    use: any;
    create: any;
  };
  subscribers: {
    subscribe: any;
    getAll: any;
  };
  reviews: {
    getByProduct: any;
    create: any;
    getAll: any;
  };
  contactMessages: {
    create: any;
    getAll: any;
  };
  wishlists: {
    get: any;
    add: any;
    remove: any;
  };
  printfulEvents: {
    create: any;
    getByOrderId: any;
  };
  categories: {
    getAll: any;
    getBySlug: any;
    create: any;
    update: any;
    remove: any;
    getWithProductCount: any;
  };
  collections: {
    getAll: any;
    getBySlug: any;
    create: any;
    update: any;
    remove: any;
  };
  seed: {
    seedProducts: any;
    clearProducts: any;
  };
};

export { api };